local/requirements.txt
pandas
numpy
python-dotenv
PyMuPDF
ipython
scikit-learn
langchain-pinecone
langchain 
langchain-community
langchain-experimental 
xmltodict
chromadb==0.5.0
langchain-google-genai
pypdf
pinecone-client
semanticscholar
langchainhub
lark
langchain-chroma>=0.1.2
google-ai-generativelanguage
langchainhub 
google-generativeai
langchain-google-community
beautifulsoup4
html2text
faiss-cpu
streamlit-authenticator
openpyxl
xlrd
langchain-elasticsearch
tabulate
detectaicore
langchain-nvidia-ai-endpoints 
markdownify 
sentence-transformers
pymilvus
langchain_milvus
langchain-google-vertexai
rapidocr-onnxruntime
pymupdf4llm
langchain-nomic
tiktoken   
langgraph 
tavily-python 
gpt4all 
langchain-text-splitters
ollama
langchain-huggingface
